requirements.txt
numpy>=1.26.4
matplotlib>=3.9.0
pillow>=10.3.0